requirements.txt
certifi==2019.6.16
chardet==3.0.4
configparser==3.7.4
Django==2.2.3
idna==2.8
mysqlclient==1.4.2.post1
python-google-places==1.4.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
zomathon==1.4
